pyproject.toml
[project]
name = "inferwall"
version = "0.1.0-dev"
requires-python = ">=3.12"
description = "Signature-based AI application firewall"
license = { text = "Apache-2.0" }
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "mypy>=1.11",
    "maturin>=1.0",
    "pre-commit>=3.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
inferwall = "inferwall.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/inferwall"]

[tool.ruff]
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["inferwall_core.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["faiss", "llama_cpp", "onnxruntime"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: fast unit tests",
    "integration: tests requiring compiled Rust extension",
    "security: adversarial/security tests",
]
